🎨(parsers) remove filters and data parsing
Parsers should extract raw events from a container format, e.g. the
short message content from a GELF-formatted log. Previous implementation
required the log message to be a JSON string and this was a mistake:
this is not the parser's reponsibility to parse the event itself but
rather the container format.