👷(circle) authenticate to pull images from DockerHub
Docker will soon add a rate limit on image pulls from DockerHub. This
rate limit is based on the request IP for anonymous users; in a CI
context, CircleCI IPs will be taken into account, hence quota will be
reached quickly.

To prevent this behavior, we should authenticate to DockerHub before
pulling or building images so that this rate limit is now on a per-user
basis.