🚀(project) add Arnold tray
Make Ralph deployable as an Arnold tray!

By default, we scale Ralph DC at zero as it is not intended to work as a
server but as a CLI. The DC could be usefull to test or experiment with
the tool. The core part of its usage should be defined as pipelines
executed in cronjobs.