✨(tray) add support for ES cluster CA certs
When requesting an elasticsearch cluster using self-generated CA
certificates, the elasticsearch client is raising a warning message
as it cannot verify the authenticity of SSL certificates.

To mitigate the issue, one need to copy the CA certificate in ralph's
container and specify the CA certificate path for the "es" backend using
the "--es-client-options ca_certs=/path/to/ca.pem" option.

If the `ES_CA_CERTIFICATE` is defined in ralph's vault, a dedicated
secret will be generated and can be mounted in the following path
/usr/local/share/ca-certificates/es-cluster.pem so that execution logs
are not polluted with warning messages.