➕(api) add fastapi and gunicorn/uvicorn to our dependencies
We want to build an API that implements xAPI to receive requests
as if it was an LRS. The first thing we need is a web server with
an API building framework.

We settled on fastapi and uvicorn to get a more performant,
asynchronous server, and perform type validation of requests.