✨(api) authenticate users through middleware
Ralph does not provide any persistence mechanism (either file-based
or a database), but still needs some form of authentication &
authorization for users of its API.

We can achieve this with HTTP basic auth using a list of user
credentials and permissions set up in a static text file.

We implemented this through a Starlette/FastAPI middleware so it
can be used transparently with every API route we open.