🐛(bin) fix double port binding causing errors
As we added the `runserver` command, we added port bindings on
`run` through the `--service-ports` option. This caused issues
when starting more than one `ralph` command (eg. for piping)
because both processes attempted to bind the same port.

We can bind ports only when running the `runserver` commands to
avoid these conflicts.