♻️(api) rework authentication to use a fastapi dependency
As we built authentication into our Ralph API, we looked for
auth middleware and found built-in, compatible middleware from
Startlette that was compatible with FastAPI.

We did not know at that time that there are incompatibilities
with this specific middleware that prevent us from doing any sort
of authorization work with FastAPI, like this middleware does if
used directly with Starlette.

We moved on to refactor our authentication to use a FastAPI
dependency instead of a middleware.