✨(elasticsearch) build a quick partial implementation of LRS GET
We're quickly building a partial implementation of the LRS
specification to experiment with customers. With the additional
constraint of not being able to build our own index for search
purposes, we're just doing what we can do with the basic existing
shape of the documents.

We're documenting all parameters for GET, specifying which ones
are actually supported or not in the current version.