♻️(tests) use a fixture for auth in statements GET tests
As we initially built the statements GET endpoint and set up its
tests, we created a function to generate authentication credentials.
When we copied that function in the POST tests, we decided to
replace it with a pytest fixture instead.

Now that we have a fixture to do this, we can remove the original
function and just use the fixture for GET statements tests too.