🎨(tests) use a custom hypothesis build strategy instead of st.builds
The hypothesis st.builds strategy always passes the value `None` for
optional fields having a default value.
However, for some of our optional Model fields the value `None` is not
valid. In other words, they should be either omitted or have a value
different from `None`.
Therefore we add a custom build strategy to overcome that.

We also add a global hypothesis configuration profile to avoid
repeating the same hypothesis settings for each test.

And we patch the hypothesis Json strategy because the default one
raises an exception during model generation.