✨(backends) allow database backends custom get queries
We need more flexibility to fetch learning events from database
backends. Allow backend-specific-queries for filtering and aggregations
seems required to achieve such task.

Each database backend `get` method now supports an optionnal `query`
field that will be injected in the backend's request. This `query` field
accepts a query model instance as an argument. Each backend is
responsible to define its own model, and check that passed arguments are
valid and respect backend-specific API.

Consequently, the `fetch` command now supports a new `-q` or `--query`
option with a JSON string argument. As query objects may be complex for
some backends, we haven't found a more user-friendly alternative for
this option in a CLI context.