🚨(backends) fix pylint warning
The get method of `requests` needs a timeout defined. However, it is not
necessary in the use of this method in the ldp backend read method
implementation. The use of pydantic models as a dict key raises an hash
warning, as no hash function has been defined in the model class.