🚸(project) improve installation flavors consistency
When Ralph is installed as a library, the `ralph` command is installed
and available in the PATH even if Click is not installed leading to
unobvious error message when executing the aforementioned command.

More than that, if installed as a CLI, the `runserver` command will fail
if the the LRS flavor is not installed.

For a better user/developer experience, we now explicitly raise an
appropriate error message on those two cases.