💥(cli) add mandatory "agent" field to credentials
In link with #288 , we need a mechanism to assign an "Agent" (in the sense of
the xAPI specification) to a user (to later be able to infer the "Authority"
field when writing a statement). It was proposed to add the "agent"
representation as a field in the user credentials, which is what is implemented
here. There is no contraint as to which of the 4 valid IFI's must be used.

BREAKING: New users must now include an agent field