✨(api) add "mine" option to fetch statements with matching authority
In link with #288 , we need a mechanism to fetch statements by authority.
It has been decided that the GET /statements endpoint will implement a mine
option (not in xAPI specification), which is used to return only the
statements for which the "authority" field matches the "agent" field of
the current user (stored in credentials since #342, or given through token
claims).
With future implementation of permissions, it is expected that the use of mine
option will be made mandatory for users of the API that do not have full read
access. The idea behind this is to make any filtering obvious to the user
("explicit is better than implicit"). This commit adds mine by fetching user
"agent" field and querying using a new database query field.