✨(backends) create async LRS backend and derive sync LRS from it
This PR aims to create an asynchronous version of the backend (largely
identical to the LRS backend implemented in #289). The (non synchronous)
LRS Backend will be implemented as an class that inherits from the async
backend, to minimize code duplication. There is a slight overhead due to
the use of asyncio, but this should be negligable compare to the waiting
time linked to API calls. It was decided: 1. to add a `greedy` option to
`read` which will automatically fetch all result pages, even before results
are consumed. 2. a `simultaneous` option to `write` to POST all chunks
simultaneously.