🦺(models) make ProfileActivity presence mandatory
The latest version of the xapi video profile (1.0.3) seems to enforce
the presence of a video profile activity in the
`context.contextActivities.category` field which was previously
validated to optionally contain the profile activity.
value. Therefore we choose to make this field presence mandatory
in statements following the video and virtual classroom profiles.