♻️(backends) remove ClickHouse JSON column (#482)
Initial implementation of the ClickHouse backend used the experimental JSON
column type, with the string of the JSON as a backup. The JSON type has never
graduated out of experimental status and has some serious issues that make it
undesirable to support.

This commit removes all references to the column in favor of using the JSON
ClickHouse functions, which are generally faster and better supported.