🚨(project) fix Ruff warnings/errors
- gitlint requests an emoji file hosted on GitHub without a timeout. (S113)
Adding a 5-second timeout.
- Docstring of `get_oidc_user` was missing a description for argument
`security_scopes`. (D417) Adding it.
- Using map function with lambda is slower and less readable than with a
comprehension. (C417) Changing them to comprehensions.
- change conversion to tuple (C409)
- fix docstring subsections (D214, D410)
- fix docstring summary on two lines (D205)
- remove lru-caching of a `UserScopes` method (B019). Every instance of
`UserScopes` will be kept alive as long as the method is cached (with a
`max_size` of 1024, they will never be garbage collected until cache cleared
or app shut down). It seems unsafe, removing the cache on this method for now.
- fix unsorted/unformatted imports (I001)