♻️(backends) remove custom validation for backend queries
A validation mechanism used to support backend queries either in query string or
based on a Pydantic model when query to a backend is custom. To simplify this
mechanism, backends only support either string typing when no query model is
specified and a the query Pydantic model in the other case.