♻️(backends) add logger property to data/lrs backends
Currently, backends use a module level logger instance
for logging to allow filtering logs by backend.
However, when a backend inherits another backend, logs
produced by the parent backend refer to the parent module.
Thus, to allow a child backend to overwrite the logger
of it's parent, we choose to make the logger an instance
property of the backend class.