♻️(backends) fix backend type hints and refactor common settings
The `DEFAULT_CHUNK_SIZE` and `LOCALE_ENCODING` settings are shared
among data backends. Thus, instead of redefining them for each
backend, we choose to move these settings to `BaseDataBackendSettings`.

The async data backend read method was wrongly typed to return an
Iterator. As it's an async generator it should return an AsyncIterator.