🐛(backends) make ldp backend raise an exception when raw_output=False
Currently, the `LDPDataBackend.read` method ignores the `raw_output`
argument and always yields bytes.
This behavior is problematic if the caller relies on the backend
to return dictionaries when `raw_output` is set to `False`.
Thus, we choose to instead raise a `BackendParameterException`
in this case.
We also wrap more API calls in try/except blocks that were previosly
omitted.