🏗️(backends) migrate LRSHTTPBackend to LRSDataBackend
The interface of the http backends are now very similar
to the data backends.
Thus, to limit redundant interfaces we opt to move the
LRSHTTPBackend and AsyncLRSHTTPBackend to the data
backend package.
We also update the cli write command to align more with
the data backend interface.
Moreover, the LRSDataBackend no longer inherits from
the AsyncLRSDataBackend to alow it's usage in an async
context.