✨(cli) replace unused archive argument with query
The `archive` argument is not used anymore, thus we replace it with
a query argument that accepts a json string or a plain string and
remove the previously defined query option which accepted json only.
We want to accept a plain string as the query argument to simplify
CLI usage with backends that don't need a dictionary query
(fs/s3/swift). Also, this simplifies the use the ES data backend with
a Lucene query string - e.g.: `ralph read -b es id:1`

We also drop the `BaseQuery.query_string` field and opt to allow
backends to have a string query_class as some backends do not
need a full query model (FS, LDP, SWIFT, S3).
Finaly, we add a dedicated `from_sting` method
in the `BaseQuery` to parse JSON string queries for backends
that use a Query model.