💚(circle) allow executing jobs for pre-release tags
We want to build and push `ralph-malph` package to PyPi for a pre-release.
We also want to build and push Ralph docker image for pre-releases.
We should be able to do so when tagging appropriately (e.g. `4.0.0-b1`).
This prevents executing the other jobs triggered on tags.
It also prevents tagging Ralph docker image as `latest` when triggered on tag.