✨(cli) make the -b/--backend option a CLI sub-command
With the addition of more backends, the usage documentation of
the ralph list/read/write and runserver commands became cluttered
as we list all available backends with all their options.
Thus, to keep the documentation concise, we choose to remove the
-b/--backend option and use a backend sub-command instead.
This allows us to have dedicated usage documentation for each backend.
We also improve backend documentation by parsing method and settings
docstrings to construct option help strings.