💚(circle) test helm with ralph build image
In the `test-helm` job, we were building Ralph current image but were installing
the helm chart without specifying an image tag. When not specified, Helm
installs the `appVersion` of the image, fetching it from DockerHub repository
and not testing the current build of the image.
Changing it so that the job tests the current build of the image for non-tagged
workflows, but also tests that the `appVersion` has correctly been filled in for
`-helm` tagged workflows.