♻️(models) change default value "Agent" object type
Until now, "Agent" object type was set to None as default value. When using
ralph xapi models to generate statements instances, objectType Agent is
generated with None value which is not the required behavior. We keep this field
as optional in Pydantic model for validation and force "Agent" value as a
default value.